Moved style and font code to interface.helpers
